+++ besturingseenheid/besturingseenheid/Ultrasonic_sensor.h
@@ -10,9 +10,9 @@
 uint16_t calc_cm(uint16_t counter);
 void update_afstand(void);
 uint16_t Get_huidige_afstand(void);
-void Set_uitrol_afstand(int uitrol);
-void Set_oprol_afstand(int oprol);
-int Get_uitrol_afstand(void);
-int Get_oprol_afstand(void);
+void Set_uitrol_afstand(uint8_t uitrol);
+void Set_oprol_afstand(uint8_t oprol);
+uint8_t Get_uitrol_afstand(void);
+uint8_t Get_oprol_afstand(void);
 
 #endif /* ULTRASONIC_SENSOR_H_ */
